# member-service/swagger/member-swagger-spec.json
{
  "swagger": "2.0",
  "info": {
    "title": "Member Service API",
    "version": "1.0.0",
    "description": "Dynamic Swagger generation for Member Service"
  },
  "host": "localhost:8001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/member": {
      "post": {
        "summary": "Create a new member",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Member object to create",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John Doe"
                },
                "age": {
                  "type": "integer",
                  "example": 30
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Member created",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all members",
        "responses": {
          "200": {
            "description": "A list of members",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "age": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/member/{id}": {
      "put": {
        "summary": "Update a member by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "The ID of the member to update"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Member object with updated details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "age": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member updated",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a member by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "The ID of the member to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Member deleted",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ping": {
      "get": {
        "summary": "Responds with pong",
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {},
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}